Search and replace all instances of unsigned with size_t. -- Test by compiling under Linux 64, Win32, and Win64. -- Win64 found the most errors because their "size_t" is not a "long". -- Required modifying NetCDF-3 C++ interface to resolve issues on Win66. -- Fix errors in unsigned math for az-1.

+++ netcdf/netcdf_profile.cc
@@ -151,10 +151,10 @@ netcdf_profile::netcdf_profile(
  */
 void netcdf_profile::fill_missing() {
 
-    const int alt_num = this->_axis[0]->size() ;
-    const int lat_num = this->_axis[1]->size() ;
-    const int lng_num = this->_axis[2]->size() ;
-    unsigned index[3] ;
+    const size_t alt_num = this->_axis[0]->size() ;
+    const size_t lat_num = this->_axis[1]->size() ;
+    const size_t lng_num = this->_axis[2]->size() ;
+    size_t index[3] ;
 
     // compute average value at each depth
 
@@ -185,7 +185,7 @@ void netcdf_profile::fill_missing() {
             if ( index[0] <= 0 ) {
                 average.data( index, NAN ) ;
             } else {
-                unsigned prev_index[1] ;
+                size_t prev_index[1] ;
                 prev_index[0] = index[0] - 1 ;
                 average.data( index, average.data(prev_index) ) ;
             }
